mvp/backend/requirements.txt
PyPDF2
uvicorn
fastapi
python-multipart
sentence-transformers
faiss-cpu
openai
python-dotenv